♻️ :: 의존성 에러 해결
